feat(client): add sentry logging
Enable sentry for capturing exceptions.

Use the telemetry config switches that the LSP uses.

Redirect formatting exceptions to the error telemetry log (and hence on
to sentry).

Part of #102 (HHVSC-124).
