+++ .changeset/heavy-lamps-attack.md
@@ -2,4 +2,4 @@
 "hardhat-solidity": patch
 ---
 
-Fix - don't throw on hh error objects with no source location
+Fix - don't throw on hh error objects with no source location ((#147)[https://github.com/NomicFoundation/hardhat-vscode/issues/147])
