+++ client/.eslintrc.js
@@ -2,8 +2,4 @@
 module.exports = {
   root: true,
   extends: [`../.eslintrc.js`],
-  parserOptions: {
-    project: `./tsconfig.json`,
-    sourceType: "module",
-  },
 };
