+++ client/package.json
@@ -13,11 +13,13 @@
     "clean": "rimraf out tsconfig.tsbuildinfo"
   },
   "devDependencies": {
+    "@sentry/types": "6.19.1",
     "@types/vscode": "^1.63.0",
     "eslint": "^7.23.0",
     "rimraf": "3.0.2"
   },
   "dependencies": {
+    "@sentry/node": "6.19.1",
     "prettier": "2.5.1",
     "prettier-plugin-solidity": "^1.0.0-beta.19",
     "vscode-languageclient": "^7.0.0"
