+++ client/src/extension.ts
@@ -1,411 +1,47 @@
-import * as path from "path";
-import * as events from "events";
-import {
-  workspace,
-  window,
-  languages,
-  ExtensionContext,
-  TextDocument,
-  OutputChannel,
-  WorkspaceFolder,
-  Uri,
-  ProgressLocation,
-  TextEdit,
-  extensions,
-  env,
-  ConfigurationTarget,
-} from "vscode";
-import {
-  Disposable,
-  LanguageClient,
-  LanguageClientOptions,
-  ServerOptions,
-  TransportKind,
-} from "vscode-languageclient/node";
+import { ExtensionContext } from "vscode";
+import { warnOnOtherSolidityExtensions } from "./popups/warnOnOtherSolidityExtensions";
+import { setupExtensionState } from "./setup/setupExtensionState";
+import { setupFormatterHook } from "./setup/setupFormatterHook";
+import { setupLanguageServerHooks } from "./setup/setupLanguageServerHooks";
+import { ExtensionState } from "./types";
 
-import { formatDocument } from "./formatter";
-import { Logger } from "./Logger";
-import { HardhatVSCodeConfig } from "./types";
+let extensionState: ExtensionState | null = null;
 
-const CONFLICTING_EXTENSION_ID = "juanblanco.solidity";
-const CONFLICTING_EXTENSION_NAME = "solidity";
-
-type IndexFileData = {
-  path: string;
-  current: number;
-  total: number;
-};
-
-const clients: Map<string, LanguageClient> = new Map();
-const listenerDisposables: Disposable[] = [];
-
-let _sortedWorkspaceFolders: string[] | undefined;
-function sortedWorkspaceFolders(): string[] {
-  if (_sortedWorkspaceFolders === void 0) {
-    _sortedWorkspaceFolders = workspace.workspaceFolders
-      ? workspace.workspaceFolders
-          .map((folder) => {
-            let result = folder.uri.toString();
-
-            if (result.charAt(result.length - 1) !== "/") {
-              result = result + "/";
-            }
-
-            return result;
-          })
-          .sort((a, b) => {
-            return a.length - b.length;
-          })
-      : [];
-  }
-
-  return _sortedWorkspaceFolders;
-}
-
-workspace.onDidChangeWorkspaceFolders(
-  () => (_sortedWorkspaceFolders = undefined)
-);
-
-function getOuterMostWorkspaceFolder(folder: WorkspaceFolder): WorkspaceFolder {
-  const sorted = sortedWorkspaceFolders();
-
-  for (const element of sorted) {
-    let uri = folder.uri.toString();
-
-    if (uri.charAt(uri.length - 1) !== "/") {
-      uri = uri + "/";
-    }
-
-    if (uri.startsWith(element)) {
-      return workspace.getWorkspaceFolder(Uri.parse(element));
-    }
-  }
-
-  return folder;
-}
-
-function sleep(ms) {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-function getUnsavedDocuments(): TextDocument[] {
-  return workspace.textDocuments.filter((i) => i.isDirty);
-}
-
-/**
- * Sends a no-op change notification to the server, this allows the
- * triggering of validation.
- * @param client the language client
- * @param textDoc the open text file to trigger validation on
- */
-function notifyOfNoopChange(client: LanguageClient, textDoc: TextDocument) {
-  client.sendNotification("textDocument/didChange", {
-    textDocument: {
-      version: textDoc.version,
-      uri: textDoc.uri.toString(),
-    },
-    contentChanges: [
-      {
-        range: {
-          start: { line: 0, character: 0 },
-          end: { line: 0, character: 0 },
-        },
-        rangeLength: 1,
-        text: "",
-      },
-    ],
-  });
-}
-
-/**
- * If the doc is open, trigger a noop change on the server to start validation.
- */
-function triggerValidationForOpenDoc(client: LanguageClient, path: string) {
-  const textDoc = workspace.textDocuments.find((d) => d.uri.path === path);
-
-  if (!textDoc) {
-    return;
-  }
-
-  notifyOfNoopChange(client, textDoc);
-}
-
-function showFileIndexingProgress(client: LanguageClient): void {
-  const em = new events.EventEmitter();
-
-  client.onReady().then(() => {
-    client.onNotification("custom/indexing-file", (data: IndexFileData) => {
-      em.emit("indexing-file", data);
-    });
-  });
-
-  // Progress bar
-  window.withProgress(
-    {
-      cancellable: true,
-      location: ProgressLocation.Notification,
-      title: "Indexing Project",
-    },
-    async (progress) => {
-      progress.report({
-        increment: 0,
-        message: "Start indexing...",
-      });
-
-      const promise = new Promise<void>((resolve) => {
-        em.on("indexing-file", (data: IndexFileData) => {
-          progress.report({
-            increment: Math.round(data.total / data.current),
-            message: `Indexing ${data.path}`,
-          });
-
-          // Files that were open on vscode load, will
-          // have swallowed the `didChange` event as the
-          // language server wasn't intialized yet. We
-          // revalidate open editor files after indexing
-          // to ensure warning and errors appear on startup.
-          triggerValidationForOpenDoc(client, data.path);
-
-          if (data.total === data.current) {
-            resolve();
-          }
-        });
-      });
-
-      await promise;
-
-      progress.report({
-        increment: 100,
-        message: `Project indexing is complete.`,
-      });
-
-      await sleep(3000);
-    }
-  );
-}
-
-async function warnOnOtherSolidityExtensions(logger: Logger) {
-  const conflictingExtension = extensions.getExtension(
-    CONFLICTING_EXTENSION_ID
-  );
-
-  if (conflictingExtension === undefined) {
-    return;
-  }
-
-  try {
-    await window.showWarningMessage(
-      `Both this extension and the \`${CONFLICTING_EXTENSION_NAME}\` (${CONFLICTING_EXTENSION_ID}) extension are enabled. They have conflicting functionality. Disable one of them.`,
-      "Okay"
-    );
-  } catch (err) {
-    logger.error(err);
-  }
-}
-
-async function showAnalyticsAllowPopup(
-  context: ExtensionContext
-): Promise<void> {
-  const shownTelemetryMessage = context.globalState.get<boolean>(
-    "shownTelemetryMessage"
-  );
-
-  if (shownTelemetryMessage) {
-    return;
-  }
-
-  const item = await window.showInformationMessage(
-    "Help us improve the Hardhat for Visual Studio Code extension with anonymous crash reports & basic usage data?",
-    { modal: true },
-    "Accept",
-    "Decline"
-  );
-
-  const isAccepted = item === "Accept" ? true : false;
-
-  const config = workspace.getConfiguration("hardhat");
-
-  config.update("telemetry", isAccepted, ConfigurationTarget.Global);
-
-  context.globalState.update("shownTelemetryMessage", true);
-}
+const SENTRY_DSN =
+  "https://9d1e887190db400791c77d9bb5a154fd@o385026.ingest.sentry.io/5469451";
 
 export function activate(context: ExtensionContext) {
-  const config: HardhatVSCodeConfig = {
-    env:
-      process.env.NODE_ENV === "development"
-        ? process.env.NODE_ENV
-        : "production",
-    version: context.extension.packageJSON.version,
-    name: context.extension.packageJSON.name,
-    hardhatTelemetryEnabled: workspace
-      .getConfiguration("hardhat")
-      .get<boolean>("telemetry"),
-  };
+  extensionState = setupExtensionState(context, { sentryDsn: SENTRY_DSN });
 
-  const module = context.asAbsolutePath(path.join("server", "out", "index.js"));
-
-  const outputChannel: OutputChannel = window.createOutputChannel("Hardhat");
-
-  const logger = new Logger(outputChannel);
+  const { logger } = extensionState;
 
   logger.info("Hardhat for Visual Studio Code Starting ...");
-  logger.info(`env: ${config.env}`);
+  logger.info(`env: ${extensionState.env}`);
 
-  warnOnOtherSolidityExtensions(logger);
+  warnOnOtherSolidityExtensions(extensionState);
 
-  context.subscriptions.push(
-    languages.registerDocumentFormattingEditProvider("solidity", {
-      provideDocumentFormattingEdits(document: TextDocument): TextEdit[] {
-        return formatDocument(document, context);
-      },
-    })
-  );
-
-  function didOpenTextDocument(document: TextDocument): void {
-    // We are only interested in solidity files
-    if (document.languageId !== "solidity" || document.uri.scheme !== "file") {
-      return;
-    }
-
-    const uri = document.uri;
-    let folder = workspace.getWorkspaceFolder(uri);
-
-    // Files outside a folder can't be handled. This might depend on the language.
-    // Single file languages like JSON might handle files outside the workspace folders.
-    if (!folder) {
-      return;
-    }
-
-    // If we have nested workspace folders we only start a server on the outer most workspace folder.
-    folder = getOuterMostWorkspaceFolder(folder);
-
-    if (!clients.has(folder.uri.toString())) {
-      // The debug options for the server.
-      // --inspect=6009: runs the server in Node's Inspector mode so VS Code can attach to the server for debugging.
-      const debugOptions = {
-        execArgv: ["--nolazy", `--inspect=${6009 + clients.size}`],
-      };
-
-      // If the extension is launched in debug mode then the debug server options are used.
-      // Otherwise the run options are used.
-      const serverOptions: ServerOptions = {
-        run: {
-          module,
-          transport: TransportKind.ipc,
-        },
-        debug: {
-          module,
-          transport: TransportKind.ipc,
-          options: debugOptions,
-        },
-      };
-
-      // Options to control the language client.
-      const clientOptions: LanguageClientOptions = {
-        // Register the server for solidity text documents.
-        documentSelector: [
-          { scheme: "file", language: "solidity", pattern: `**/*.sol` },
-        ],
-        diagnosticCollectionName: "hardhat-language-server",
-        workspaceFolder: folder,
-        outputChannel: outputChannel,
-        initializationOptions: {
-          extensionName: config.name,
-          extensionVersion: config.version,
-          env: config.env,
-          globalTelemetryEnabled: env.isTelemetryEnabled,
-          hardhatTelemetryEnabled: config.hardhatTelemetryEnabled,
-          machineId: env.machineId,
-        },
-      };
-
-      logger.info(`[LS: ${folder.name}] Client starting`);
-
-      // Create the language client and start the client.
-      // Start the client. This will also launch the server
-      const client = new LanguageClient(
-        "hardhat-language-server",
-        "Hardhat Language Server",
-        serverOptions,
-        clientOptions
-      );
-
-      showAnalyticsAllowPopup(context);
-
-      client.onReady().then(() => {
-        logger.info(`[LS: ${folder.name}] Client ready`);
-
-        client.onNotification("custom/get-unsaved-documents", () => {
-          const unsavedDocuments = getUnsavedDocuments();
-
-          client.sendNotification(
-            "custom/get-unsaved-documents",
-            unsavedDocuments.map((unsavedDocument) => {
-              return {
-                uri: unsavedDocument.uri,
-                languageId: unsavedDocument.languageId,
-                version: unsavedDocument.version,
-                content: unsavedDocument.getText(),
-              };
-            })
-          );
-        });
-      });
-
-      showFileIndexingProgress(client);
-
-      const telemetryChangeDisposable = env.onDidChangeTelemetryEnabled(
-        (enabled: boolean) => {
-          client.sendNotification("custom/didChangeGlobalTelemetryEnabled", {
-            enabled,
-          });
-        }
-      );
-
-      const hardhatTelemetryChangeDisposable =
-        workspace.onDidChangeConfiguration((e) => {
-          if (!e.affectsConfiguration("hardhat.telemetry")) {
-            return;
-          }
-
-          client.sendNotification("custom/didChangeHardhatTelemetryEnabled", {
-            enabled: workspace
-              .getConfiguration("hardhat")
-              .get<boolean>("telemetry"),
-          });
-        });
-
-      listenerDisposables.push(telemetryChangeDisposable);
-      listenerDisposables.push(hardhatTelemetryChangeDisposable);
-
-      client.start();
-      clients.set(folder.uri.toString(), client);
-    }
-  }
-
-  workspace.onDidOpenTextDocument(didOpenTextDocument);
-  workspace.textDocuments.forEach(didOpenTextDocument);
-  workspace.onDidChangeWorkspaceFolders((event) => {
-    for (const folder of event.removed) {
-      const client = clients.get(folder.uri.toString());
-
-      if (client) {
-        clients.delete(folder.uri.toString());
-        client.stop();
-      }
-    }
-  });
+  setupFormatterHook(extensionState);
+  setupLanguageServerHooks(extensionState);
 }
 
 export function deactivate(): Thenable<void> {
+  if (!extensionState) {
+    return;
+  }
+
+  extensionState.listenerDisposables.forEach((disposable) =>
+    disposable.dispose()
+  );
+
   const promises: Thenable<void>[] = [];
 
-  listenerDisposables.forEach((disposable) => disposable.dispose());
-
-  for (const client of clients.values()) {
+  for (const client of extensionState.clients.values()) {
     promises.push(client.stop());
   }
 
-  return Promise.all(promises).then(() => undefined);
+  const telemetryClosePromise = extensionState.telemetry.close();
+
+  return Promise.all([...promises, telemetryClosePromise]).then(
+    () => undefined
+  );
 }
