+++ client/src/telemetry/types.ts
@@ -1,0 +1,7 @@
+import { ExtensionState } from "../types";
+
+export interface Telemetry {
+  init(extensionState: ExtensionState): void;
+  captureException(err: unknown): void;
+  close(): Promise<boolean>;
+}
