+++ client/src/utils/Logger.ts
@@ -1,10 +1,13 @@
 import { OutputChannel } from "vscode";
+import { Telemetry } from "../telemetry/types";
 
 export class Logger {
   private outputChannel: OutputChannel;
+  private telemetry: Telemetry;
 
-  constructor(outputChannel: OutputChannel) {
+  constructor(outputChannel: OutputChannel, telemetry: Telemetry) {
     this.outputChannel = outputChannel;
+    this.telemetry = telemetry;
   }
 
   log(text: string) {
@@ -18,6 +21,8 @@ export class Logger {
   }
 
   error(err: unknown) {
+    this.telemetry.captureException(err);
+
     const message = err instanceof Error ? err.message : String(err);
 
     this.outputChannel.appendLine(
