+++ client/src/utils/getUnsavedDocuments.ts
@@ -1,0 +1,5 @@
+import { workspace, TextDocument } from "vscode";
+
+export function getUnsavedDocuments(): TextDocument[] {
+  return workspace.textDocuments.filter((i) => i.isDirty);
+}
